Inbuilt models

Inbuilt models

Most users of SeisModels will probably want to simply calculate properties for one of the inbuilt models, which are all exported. These are listed below:

Earth models

SeisModels.Earth.AK135 — Constant.

AK135

1-D layered model of the Earth, containing density.

Reference

Kennett, B.L.N., Engdahl, E., Buland, R., 1995. Constraints on seismic velocities in the Earth from travel-times. Geophys J Int 122, 108–124. https://doi.org/10.1111/j.1365-246X.1995.tb03540.x

source

SeisModels.Earth.EK137 — Constant.

EK137

1-D layered model of the Earth, containing density.

Reference

Kennett, B.L.N., 2020. Radial earth models revisited. Geophysical Journal International 222, 2189–2204. https://doi.org/10.1093/gji/ggaa298

source

SeisModels.Earth.IASP91 — Constant.

IASP91

1-D model of velocity.

N.B. iasp91 does not contain density.

This version of the model uses the original coefficients and is a PREMPolyModel as published. Other implementations available may use a layered piecewise linear model and/or add density from other sources.

Reference

Kennett B.L.N. Engdahl E. 1991. Traveltimes for global earthquake location and phase identification. Geophys J Int 105 429–465. https://doi.org/10.1111/j.1365-246X.1991.tb06724.x

source

SeisModels.Earth.PREM — Constant.

PREM

1-D model containing density, radial anisotropy and attenuation.

Reference

Dziewoński, A., Anderson, D., 1981. Preliminary reference Earth model. Phys. Earth Planet. Inter. 25, 297–356. https://doi.org/10.1016/0031-9201(81)90046-7

source

SeisModels.Earth.PREM_NOOCEAN — Constant.

PREM_NOOCEAN

Version of PREM where the liquid ocean layer is replaced by the crust beneath. This model may be useful when the surface must be a solid.

source

SeisModels.Earth.STW105 — Constant.

STW105

1-D layered model of the Earth, containing radial anisotropy, density and attenuation.

Attenuation is taken from QL6 (Durek and Ekström, 1996).

This version was obtained from: https://ds.iris.edu/ds/products/emc-stw105/ The original model is parameterised by radial splines.

Reference

Kustowski, B., Ekström, G., Dziewoński, A., 2008. Anisotropic shear-wave velocity structure of the Earth’s mantle: A global model. J Geophys Res-Sol Ea 113, B06306. https://doi.org/10.1029/2007JB005169

source

Moon models

SeisModels.Moon.MOON_WEBER_2011 — Constant.

MOON_WEBER_2011

1-D layered model of the Moon, containing density.

Reference

Weber, R.C.,, Lin, P.-Y.,, Garnero, E.J., Williams, Q., Lognonné, P., 2011. Seismic Detection of the Lunar Core. Science, 331, 309–312. https://doi.org/10.1126/science.1199375

source