Function index

Index of functions

Basic physical properties

SeisModels.evaluate — Function.

evaluate(m::SeisModel1D, field::Symbol, r; depth=false) -> vals

Evaluate the model m at radius r km for the different property/ies in field, returning a scalar for scalar input, and an array for array input.

If depth is true, r is treated as a depth in km instead.

source

SeisModels.vp — Function.

vp(m::SeisModel1D, r; depth=false) -> isotropic average Vp

Return the value of isotropic average Vp (km/s) for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.vs — Function.

vs(m::SeisModel1D, r; depth=false) -> isotropic average Vs

Return the value of isotropic average Vs (km/s) for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.density — Function.

density(m::SeisModel1D, r; depth=false) -> density

Return the value of density (g/cm^3) for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.vph — Function.

vph(m::SeisModel1D, r; depth=false) -> Vph

Return the value of Vph (km/s) for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.vpv — Function.

vpv(m::SeisModel1D, r; depth=false) -> Vpv

Return the value of Vpv (km/s) for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.vsh — Function.

vsh(m::SeisModel1D, r; depth=false) -> Vsh

Return the value of Vsh (km/s) for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.vsv — Function.

vsv(m::SeisModel1D, r; depth=false) -> Vsv

Return the value of Vsv (km/s) for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.eta — Function.

eta(m::SeisModel1D, r; depth=false) -> η

Return the value of η for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.Qμ — Function.

Qμ(m::SeisModel1D, r; depth=false) -> Qμ

Return the value of Qμ for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.Qmu — Function.

Qμ(m::SeisModel1D, r; depth=false) -> Qμ

Return the value of Qμ for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.Qκ — Function.

Qκ(m::SeisModel1D, r; depth=false) -> Qκ

Return the value of Qκ for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

SeisModels.Qkappa — Function.

Qκ(m::SeisModel1D, r; depth=false) -> Qκ

Return the value of Qκ for model m at radius r km.

If depth is true, then r is given as a depth in km instead.

source

Model enquiry

SeisModels.depth — Function.

depth(m, radius) -> depth

Return the depth in km for the model m given a radius in km.

source

SeisModels.hasattenuation — Function.

hasattenuation(m) -> ::Bool

Return true if the model m includes attenuation and false otherwise.

source

SeisModels.hasdensity — Function.

hasdensity(m) -> ::Bool

Return true if density is defined for model m, and false otherwise.

source

SeisModels.hasreffrequency — Function.

hasreffrequency(m::PREMPolyModel) -> ::Bool

Return true if a reference frequency is defined for PREMPolyModel m, and false otherwise.

source

SeisModels.isanisotropic — Function.

isanisotropic(m) -> ::Bool

Return true if the model m is anisotropic and false otherwise.

source

SeisModels.radius — Function.

radius(m, depth) -> radius

Return the radius in km for the model m given a depth in km.

source

SeisModels.reffrequency — Function.

reffrequency(m::PREMPolyModel) = f

Return the reference frequency for a PREMPolyModel m in Hz.

source

SeisModels.surface_radius — Function.

surface_radius(m) -> radius

Return the surface radius in km of the SeisModel m.

source

Derived model properties

SeisModels.bulk_modulus — Function.

bulk_modulus(m::SeisModel1D, r; depth=false) -> K

Return the bulk modulus K in GPa at radius r km in the model m.

If depth is true, r is treated as a depth in km instead.

source

SeisModels.gravity — Function.

gravity(m::SeisModel1D, r) -> g

Return the acceleration due to gravity, g, in m/s^2 at radius r km.

source

SeisModels.mass — Function.

mass(m::SeisModel1D, r; depth=false) -> mass

Return the mass in kg between the centre of the model and the radius r km.

If depth is true, r is treated as a depth in km instead.

source

SeisModels.moment_of_inertia — Function.

moment_of_inertia(m, r0=0, r1=surface_radius(m)) -> I

Return the moment of interia I in kg m² for the model m between radii r0 and r1 in km.

source

SeisModels.poissons_ratio — Function.

poissons_ratio(m, r; depth=false) -> ν

Return the Poisson's ratio for the model m given a radius r in km. The calculation uses the isotropic average velocities at r.

If depth is true, r is treated as a depth in km instead.

source

SeisModels.pressure — Function.

pressure(m::SeisModel1D, r; depth=false) -> p

Return the pressure p in Pa at radius r km.

If depth is true, r is treated as a depth in km instead.

source

SeisModels.shear_modulus — Function.

shear_modulus(m::SeisModel1D, r; depth=false) -> μ

Return the shear modulus μ (often also called G) in GPa at radius r in the model m.

If depth is true, r is treated as a depth in km instead.

source

SeisModels.surface_mass — Function.

surface_mass(m::SeisModel1D, r; depth=false) -> mass

Return the mass in kg betwen radius r km and the surface.

If depth is true, r is treated as a depth in km instead.

source

SeisModels.youngs_modulus — Function.

youngs_modulus(m, r; depth=false) -> E

Return the Young's modulus E in GPa for the model m given a radius r in km.

If depth is true, r is treated as a depth in km instead.

source

Input/output

SeisModels.read_mineos — Function.

read_mineos(file) -> m::LinearLayeredModel

Read a 1D seismic model from a Mineos tabular format file.

References:

https://geodynamics.org/cig/software/mineos/mineos-manual.pdf

source

SeisModels.write_mineos — Function.

write_mineos(m::LinearLayeredModel, file, freq=1.0, title="Model from SeisModels.jl")

Save a SeisModel1D as a Mineos 'tabular' format file. Supply the reference frequency freq in Hz (which defaults to 1 Hz) and a title.

Note that Mineos has two types of model file; one parameterised by radial knots with values interpolated linearly between them ('tabular' format), and one with PREM-style polynomials ('polynomial' format). Mineos uses the tabular format internally even if a polynomial file is read in, and testing suggests that polynomial files are not used correctly by the CIG version of Mineos, so it is not recommended to use them at all. Consequently, SeisModels does not support writing polynomial format files.

Reference

Mineos manual, https://geodynamics.org/cig/software/mineos/mineos-manual.pdf

source